Bring getClearLink function into queryParamsMixin
